+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Atlassian Downloader - Changelog
 
-* `1.0.0.6` - added support  of `clover`. fixed broken json parsing. added new logger.
+* `1.0.0.7` - added `unofficial support`  of `sourcetree` via automatic mirror [from github](https://github.com/EpicMorg/atlassian-json). fixed `logger` output, code improvments.
+* `1.0.0.6` - added support of `clover`. fixed broken json parsing. added new `logger`.
 * `1.0.0.5` - added support for `EAP` releases.
 * `1.0.0.4` - bump version. rewrited build scripts. added support of `arm` and `arm64`.
 * `1.0.0.3` - some cosmetics improvements.